Add the 5-round play feature
Create the playRound function and let the user
play the game for 5 rounds using a for loop

In the for loop, if the user type in invalid word,
alert them to force him to write the correct one.
Once the user input the correct word, call the
playRound()

In the playRound function, take in human choice
and computer choice and compare them to output
results. The result include: human and computer
choices, the result of the round (who win or tie),
and each player's current score